Inline useDoubleTapToZoom
